requirements.txt
python-telegram-bot==21.11.1
yt-dlp==2026.02.18
python-dotenv==1.0.1
